+++ src/main/java/com/joojoo/api/search/application/SearchServiceImpl.java
@@ -17,14 +17,13 @@ import com.joojoo.global.common.enums.SearchTarget;
 import com.joojoo.global.exception.handleException.tags.TagNotFoundException;
 import com.joojoo.global.exception.handleException.tickers.TickerNotFoundException;
 import com.joojoo.global.exception.handleException.users.UserNotFoundException;
-import com.joojoo.global.util.HangulUtils;
+import com.joojoo.global.util.hangul.HangulUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 
-import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
