+++ src/main/java/com/joojoo/api/util/infrastructure/jaso/HangulJasoAdapter.java
@@ -1,7 +1,7 @@
 package com.joojoo.api.util.infrastructure.jaso;
 
 import com.joojoo.api.util.port.in.JasoConverter;
-import com.joojoo.global.util.HangulUtils;
+import com.joojoo.global.util.hangul.HangulUtils;
 import org.springframework.stereotype.Component;
 
 @Component
