+++ src/main/java/com/joojoo/global/util/codeRandom/CodeGenerator.java
@@ -1,5 +1,6 @@
-package com.joojoo.api.util.random;
+package com.joojoo.global.util.codeRandom;
 
+import com.joojoo.api.util.port.in.GeneratorRandom;
 import org.springframework.stereotype.Service;
 
 import java.util.Random;
